Create README.md
we are a federal company that has been buying and selling new used cars for over 16 years. We are represented in 9 countries of the region and we do not want to stop there. Now we see great opportunities, and therefore we are actively developing in all areas of our business. One of these attracts is our IT products.